// src/api/generated/model/listAllPostsOrderBy.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */

export type ListAllPostsOrderBy = (typeof ListAllPostsOrderBy)[keyof typeof ListAllPostsOrderBy];

export const ListAllPostsOrderBy = {
  createdAt: "createdAt",
  likes: "likes",
  comments: "comments",
} as const;